AcetzBenchmark: Merge benchmark_zoneinfo.py into benchmark.py; add Makefile to benchmark.txt
