+++ benchmarks/AcetzBenchmark/Makefile
@@ -1,0 +1,5 @@
+benchmark.txt:
+	python3 benchmark.py > $@
+
+clean:
+	rm -f benchmark.txt
