+++ benchmarks/AcetzBenchmark/benchmark.txt
@@ -1,0 +1,11 @@
+START
+Original timezones: 377
+Common timezones: 377
+Start year: 2000
+Until year: 2038
+BENCHMARKS
+acetz 343824 10.256 349102 12.700
+dateutil 343824 5.974 349102 7.372
+pytz 343824 15.744 349102 15.286
+zoneinfo 343824 1.415 349102 0.636
+END
